blog_food/requirements/base.txt
asgiref==3.4.1
astroid==2.6.6
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.12.0
importlib-metadata==4.6.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
packaging==21.0
pep517==0.11.0
Pillow==8.3.1
pip-review==1.1.0
pipdeptree==2.1.0
pylint==2.9.6
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
yapf==0.31.0
zipp==3.5.0

